Fixes Nextcloud server v21 compatibility
End-to-End Encryption v1.6.2 (the current version available in the apps
repository) is not compatible with Nextcloud server v21 due to a
breaking change in the DBAL Types interface. This has been fixed in the
unreleased master branch, however the `appinfo/info.xml` file in that
branch lists v21 as the min-version supported of the Nextcloud server.

I have decremented min-version to v21 and tested on a local v21 server
based on the docker-compose v21 postgres deployment, as published by
nextcloud. E2EE installs and seems to work properly. I was able to setup
an encrypted folder, add some files, verify that they were (or at least
appeared to be) encrypted on the server, setup a second client, enter
the mnemonic, and read the encrypted files on the second client. I'd
hope to get this pushed to production as soon as possible to fix this
regression, but obviously given the sensitive nature, someone with more
experience should verify that the encryption does indeed work properly
on v21.