fix: remove lock without mountpoint
A folder is locked to keep file and metadata in sync[^1].

`oc_e2e_encryption_lock.id` is related to `oc_filecache.fileid`.

Is the filecache entry gone, the lock entry is in an undefined state. It's not possible to release the lock due to the missing reference, and the lock will stay forever.

[^1]: https://github.com/nextcloud/end_to_end_encryption_rfc/blob/master/RFC.md#update-metadata-file

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>