fix: Remove legacy optimization when checking encryption status
This is causing performance issues because for non encrypted nodes, we were checking all upper folders in the hierarchy.

Also, this seems unnecessary as top and sub E2EE folders are marked as encrypted, so checking for the current node or its parent folder is enough to have the answer.

Partially introduced by https://github.com/nextcloud/end_to_end_encryption/pull/316 which was an improvement at the time by introducing caching.

Signed-off-by: Louis Chemineau <louis@chmn.me>